egress/requirements.txt
alabaster==0.7.12
argcomplete==1.11.0
astroid==2.3.3
Babel==2.8.0
bcrypt==3.1.7
blockdiag==1.5.3
boto3==1.10.9
botocore==1.13.50
cachetools==4.0.0
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
cryptography==2.8
Cython==0.28.5
dill==0.2.7.1
docutils==0.15.2
Flask==1.1.1
funcparserlib==0.3.6
future==0.16.0
get==2019.4.13
glob2==0.6
google-api-core==0.1.4
google-api-python-client==1.6.5
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.7.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
h5py==2.10.0
httplib2==0.15.0
idna==2.6
imagesize==1.2.0
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.3
lxml==4.3.0
m2r==0.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
numpy==1.16.3
oauth2client==4.1.3
opencv-python-headless==4.1.0.25
packaging==20.0
paramiko==2.7.1
Pillow==6.2.0
pockets==0.9.1
post==2019.4.13
protobuf==3.6.1
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycparser==2.19
Pygments==2.5.2
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.6
pysftp==0.2.9
python-dateutil==2.7.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-toolbelt==0.8.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.19.2
scipy==0.19.1
simplejson==3.8.1
six==1.11.0
snowballstemmer==2.0.0
Sphinx==1.7.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-websupport==1.1.2
tabulate==0.8.5
typed-ast==1.4.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==16.7.9
webcolors==1.10
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
git+git://github.com/voxel51/api-py.git@develop
git+git://github.com/voxel51/eta.git@install-fix
